README.md
I am a bachelor student working on my code writing skills and preparing myself for future opportunities as you can see my field of interest is coding in games and optimizing them for a better experience. So if I can get some help related t that then it will mean a lot to me.